refactor(events): consolidar messageCreate en folder de sub-handlers
Mismo patrón que src/events/messageReactionAdd/: un único archivo
declara `createEvent({ name: "messageCreate" })` y delega a sub-handlers
hermanos. Cero cambio funcional vs el comportamiento previo de main.

Estructura nueva:
- index.ts: pre-filtros (Disboard, bots) + delegación al chain
- memes.ts: handleMemes — aditivo, agrega reacciones en CHANNELS.MEMES
- autoThread.ts: handleAutoThread — crea thread auto en canales config
- aiMention.ts: handleAiMention — mention IA con chat() y cooldown 15s
- thanks.ts: handleThanks — detección de gracias + notif reputación

Los handlers mutuamente excluyentes devuelven Promise<boolean>; el index
corta la cadena con `if (await handleX(...)) return;`. handleMemes es
void porque puede coexistir con los demás.

Además fusiona memeReactions.ts dentro de la misma cadena. Seyfert sólo
registra UN handler por evento (this.values[name] = instance), tener
dos archivos con name:"messageCreate" hacía que uno sobrescribiera
silenciosamente al otro según el orden de readdir del filesystem
(undeterministic). Bajo el patrón nuevo todo vive en un solo
createEvent y la lógica de memes corre en su propio handler aditivo.

Extra: console.log → console.error en el catch del bootstrap (index.ts)
por consistencia con el resto de catches.
