+++ src/index.ts
@@ -39,7 +39,7 @@ async function boostrap() {
 }
 
 await boostrap().catch((error) => {
-	console.log(error);
+	console.error(error);
 	process.exit(1);
 });
